chexpert/project/requirements.txt
numpy==1.16.2
matplotlib==3.0.3
scikit-learn==0.20.3
# tensorflow==1.15.4
tensorboardX==1.6
easydict==1.9
opencv-python==4.0.0.21
PyYAML
typer
torch
torchvision
pandas
tqdm
